[JSC] Optimize some of Temporal date operations https://bugs.webkit.org/show_bug.cgi?id=317091 rdar://179644689
Reviewed by Sosuke Suzuki.

This patch adds several optimizations and cleanup of date related operations.

1. Change balanceISODate -> addDaysToISODate to align to the spec. Also
   this makes code clean by getting PlainDate directly.
2. Add fast path to addDaysToISODate, which does not cross the month
   with the added days.
3. Add fast path to adjustDateDurationRecord which can quickly check the
   only given fields.
4. Add fast path to isDateTimeWithinLimits for known valid range of years.

Test: Source/JavaScriptCore/API/tests/TemporalCoreTest.cpp

* Source/JavaScriptCore/API/tests/TemporalCoreTest.cpp:
(JSC::TemporalCore::testAddDaysToISODate):
(JSC::TemporalCore::testAdjustDateDurationRecord):
(JSC::TemporalCore::testISODateLimits):
(JSC::TemporalCore::runTemporalRSTests):
(JSC::TemporalCore::testBalanceISODate): Deleted.
* Source/JavaScriptCore/runtime/ISO8601.cpp:
(JSC::ISO8601::isDateTimeWithinLimits):
* Source/JavaScriptCore/runtime/temporal/core/DurationArithmetic.cpp:
(JSC::TemporalCore::adjustDateDurationRecord):
(JSC::TemporalCore::computeNudgeWindow):
(JSC::TemporalCore::nudgeToZonedTime):
(JSC::TemporalCore::nudgeToDayOrTime):
* Source/JavaScriptCore/runtime/temporal/core/DurationArithmetic.h:
* Source/JavaScriptCore/runtime/temporal/core/ISOArithmetic.cpp:
(JSC::TemporalCore::addDaysToISODate):
(JSC::TemporalCore::isoDateAdd):
(JSC::TemporalCore::diffISODateTime):
(JSC::TemporalCore::balanceISODate): Deleted.
* Source/JavaScriptCore/runtime/temporal/core/ISOArithmetic.h:
* Source/JavaScriptCore/runtime/temporal/core/PlainDateTimeCore.cpp:
(JSC::TemporalCore::differencePlainDateTimeWithRounding):
* Source/JavaScriptCore/runtime/temporal/core/ZonedDateTimeCore.cpp:
(JSC::TemporalCore::differenceZonedDateTime):

Canonical link: https://commits.webkit.org/315205@main

+++ Source/JavaScriptCore/runtime/temporal/core/DurationArithmetic.h
@@ -114,7 +114,7 @@ constexpr int32_t unitIndexInTable(TemporalUnit);
 
 constexpr TemporalUnit unitInTable(int32_t);
 
-TemporalResult<ISO8601::Duration> JS_EXPORT_PRIVATE adjustDateDurationRecord(const ISO8601::Duration& dateDuration, double days, std::optional<double> weeks, std::optional<double> months);
+TemporalResult<ISO8601::Duration> JS_EXPORT_PRIVATE adjustDateDurationRecord(const ISO8601::Duration& dateDuration, int64_t days, std::optional<int64_t> weeks, std::optional<int64_t> months);
 
 TemporalResult<Nudged> JS_EXPORT_PRIVATE nudgeToCalendarUnit(int32_t sign,
     const ISO8601::InternalDuration&, Int128 originEpochNs, Int128 destEpochNs,
